put main methods into respective classes
